Move Index to module.
